refactor: 필요없는 파일 제거 및 파일 widgets안으로 이동
